Much more consistent casting
Task inputs are now always cast to an expected type. This is safer, and
allows for much better composability: there is no need to specify a Task
type in the inputs dictionary. Specifying a plain type instead allows any
task outputting the right type to be used (or just a value instead of a
task). Improvements also allow for better introspection of Task
dependencies (as required by networkx or snakemake interfaces)

An important exception are the scipy distribution types (which aren't
types but generator functions). I haven't figured out how to treat these
yet, so for now the code aborts whenever it sees such a type and trusts
that the user provided the correct type.