WIP: Don't require tuples as return values. Better typing. Smoke test
The requirement to wrap outputs with tuples was confusing and a cause of 
constant bugs. With better use of typing, I think we can do without.

This commit implements basic support for Pydantic-style types with 
validators. The goal is to deprecate our custom type-casting, and use 
the same methods as Pydantic. 

'example' moved to 'examples'

Smoke test is a batch script which runs each example.