Don't unnecessarily convert nested BaseModel objects to dict in run()
Beyond being CPU wasteful, if an input has a type which is a subclass of 
the one specified in the Task, it may not be possible to reconstruct it 
without special casing (e.g. parsing the dict and determining which 
subclass to use to reconstruct). This is especially problematic mith 
sinn Models, because each Model has a different signature.