Add UnpureMemoizedTask
This task was added to address the issue of dependency on on-disk data.
The solution is to represent the Task in terms of its name and input (as 
with other tasks), but to compute its digest based on its output. This 
way, adding and removing data will trigger a new computation of a 
workflow.
As a consequence of this approach, the Task *must* be memoized, to avoid 
computing it every time we query its digest.