More reliable hashed digests.
Because digests are stored with the model to make them robust to 
modifications of inputs, they were being hashed when computing the 
digest.
This was especially problematic with nested Tasks, since `.dict()`'s 
'exclude' keyword does not operate recursively (it needs to be specified 
separately for each nested model).
This commit solves the problem by dict constructing the dictionary 
itself rather than relying on Pydantic. It recursively calls 
`compute_hashed_digest` on any Task dependencies.