Small fixes. Add `get_field_value` option to view.config
`get_field_values` allows setting a custom function for retrieving 
values in a Record. This is necessary because equivalent values are not 
always recorded the same way. For instance, task inputs may be specified 
as other tasks or as precomputed values, and those are recorded 
differently.
By default `get_field_value` maps to `utils.get_task_param`, which deals 
with some simple cases.