Allow json_encoders to be modified up until task creation
Previously, the value of `json_encoders` was fixed when smttask was 
loaded. This prevented client libraries from adding their custom types 
(the only way was to update the upstream dependency 
mackelab-toolbox.typing, before smttask was loaded).
Now the value of smttask's `json_encoders` is re-read when a task is 
created.
Clients must still take care that their import order is such that this 
this value is updated before creating.
Allowing completely arbitrary import order, beyond breaking immutability 
of Task definition, would be a lot more complicated to implement 
cleanly.