[ui] Add `create_surrogates` command
This command is used to create surrogate records when tasks have already 
been run (there output can be found in the data store) but no associated 
entry exists in the record store. The surrogate records allow all output 
files to be found via the record store, with associated task parameters.