Bugfix, `Create`: Use the current dict of json encoders when creating a task
Also add option to pass additional json encoders to `Create`, as we allow
for Task decorators.
Previously, the dict of encoders that was used was that one which was
current at the time the task_generators module is imported. Any types
later added to the dict were thus non-serializable by Tasks created with
`Create`, in contrast to other tasks.