Change MemoizedTask to always use `cache=True` as default
The name "MemoizedTask" strongly suggests that task execution will be, 
well, memoized – i.e. cached.
Previously, if `cache` wasn't set explicitely for a particular task, the 
default would be to do the same as `RecordedTask` and look at 
`config.cache_runs`. My own experience suggests that this is unexpected, 
and also rarely useful: if a Task result is so large that I worry that 
it will blow up my memory, I will likely make it a RecordedTask so that 
it is saved to disk.