[examples] Add importable 'examples' module
Sometimes when testing a worflow, we just need any task instance. Since 
production tasks often have non-trivial signatures with custom input 
types, it becomes tedious to construct correct signatures when all we 
care about is that the task exists, not what it does. The 'examples' 
module provides lightweight tasks which can be used as placeholders.