Add support for scityping annex files. Minor clean up workflows.py
When executing tasks, the scityping annex_directory is set to the task’s
output directory. That way scityping to write annex binary files alongside
the task’s main output file. The additional annex files are recorded
to the Sumatra run record, and are automatically retrieved when we reload
the main result file(s) for a task.

During testing some issues arose with the functionality in workflows;
these were also fixed, so that all tests now pass.