BugFix: Remove collisions in task instance dict
The `taskdict` registry keeps track of all instantiated tasks,
so that the same tasks is not instantiated twice with the same inputs.
However only the inputs were used to create the registry key.
So if two tasks used the same inputs, the second would not be created,
instead returning the instance of the first.

This is fixed by using both the task name and inputs for the registry key.