Provide xxhash as hashing lib and make it default
Currently users can choose between the older hash functions using
`hashlib.sha1` from the standard library, or the functions from the
third-party library `xxhash`.

I am not sure of the value in not using `xxhash` though, other than
avoiding a dependency; it’s much better aligned with what we use it for:
non-cryptographic, focus on speed, stability, and hash collisions.